# Optimize view_as_complex and view_as_real